+++ CHANGELOG.md
@@ -39,6 +39,10 @@ Converting from v3.x to v4.0
 - We created a new unified command line interface that will replace the
   single-tool command line executables (`scale`, `ik`, `id`, `rra`, `cmc`,
   etc.).
+  - `scale -S setup.xml` -> `opensim run-tool setup.xml`.
+  - `scale -PS` -> `opensim print-xml scale`
+  - `scale -PropertyInfo ...` -> `opensim info ...`
+  - `versionUpdate ...` -> `opensim update-file ...`
 
 Composing a Component from other components
 -------------------------------------------